+++ .changeset/thin-windows-attend.md
@@ -1,0 +1,5 @@
+---
+'@coinbase/onchainkit': patch
+---
+
+Adding `getFarcasterUserAddress` utility to extract user's custody or verified address. This PR also updates the version of `@types/jest` package
