+++ src/farcaster/types.ts
@@ -1,19 +1,19 @@
 /**
- * Frame address type to determine if the client is requesingt  verified addresses or a custody address
+ * Frame address type to determine if the client is requesingt  verified address or a custody address
  *  Note: exported as public Type
  */
 
 export enum FarcasterAddressType {
-  VerifiedAddresses,
+  VerifiedAddress,
   CustodyAddress,
 }
 
 /**
- * GetFarcasterUserAddressesRequest
+ * getFarcasterUserAddressRequest
  *
  * Note: exported as public Type
  */
-export interface GetFarcasterUserAddressesRequest {
+export interface getFarcasterUserAddressRequest {
   fid: number;
   farcasterAddressType: FarcasterAddressType;
 }
