--- src/utils/neynar/user/neynarGetCustodyAddressForFid.ts
@@ -1,30 +1,0 @@
-import { version } from '../../../version';
-import { FetchError } from '../exceptions/FetchError';
-import { NEYNAR_DEFAULT_API_KEY } from '../frame/neynarFrameFunctions';
-
-export async function neynarGetCustodyAddressesForFid(
-  fid: number,
-  apiKey: string = NEYNAR_DEFAULT_API_KEY,
-): Promise<string> {
-  const options = {
-    method: 'GET',
-    url: `https://api.neynar.com/v1/farcaster/custody-address?fid=${fid}`,
-    headers: {
-      accept: 'application/json',
-      api_key: apiKey,
-      'content-type': 'application/json',
-      onchainkit_version: version,
-    },
-  };
-  const resp = await fetch(options.url, options);
-  if (resp.status !== 200) {
-    throw new FetchError(`non-200 status returned from neynar : ${resp.status}`);
-  }
-  const responseBody = await resp.json();
-
-  if (!responseBody || !responseBody.result || !responseBody.result.custodyAddress) {
-    throw new Error('No custody address found for FID ' + fid);
-  }
-
-  return responseBody.result.custodyAddress;
-}
